depgraph: Add trace-internal command
